# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.envelope:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.errorEnvelope:
    properties:
      errors:
        items:
          $ref: '#/definitions/handler.fieldError'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.fieldError:
    properties:
      detail:
        type: string
      field:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://github.com/404nfidv2/go-nuxt-starter-kit
  description: REST API for the Go-Nuxt Starter Kit (auth, profile, notifications,
    admin).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go-Nuxt Starter Kit API
  version: "1.0"
paths:
  /admin/announcements:
    post:
      consumes:
      - application/json
      parameters:
      - description: Announcement
        in: body
        name: body
        required: true
        schema:
          properties:
            body:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Broadcast a system announcement to all active users via WebSocket
      tags:
      - admin
  /admin/users:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by email verification
        in: query
        name: is_email_verified
        type: boolean
      - description: Search by name or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: List all users (paginated, filterable)
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: New user details
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Create a user (sends set-password email)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Hard-delete a user immediately
      tags:
      - admin
    get:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - admin
    patch:
      consumes:
      - application/json
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        schema:
          properties:
            bio:
              type: string
            display_name:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            role:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Update a user's profile or status
      tags:
      - admin
  /admin/users/{id}/activate:
    post:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Set user status to active
      tags:
      - admin
  /admin/users/{id}/ban:
    post:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Ban a user
      tags:
      - admin
  /admin/users/{id}/deactivate:
    post:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Set user status to inactive
      tags:
      - admin
  /admin/users/{id}/unban:
    post:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Unban a user (sets status to active)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"